requirements.txt
runpod
python-dotenv
langchain
langchain_community
huggingface_hub
text-generation
